% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clowder_auth.R
\name{clowder_auth}
\alias{clowder_auth}
\title{Setup authentication parameters for connecting to clowder}
\usage{
clowder_auth(user = NULL, pass = NULL, apikey = NULL, baseurl)
}
\arguments{
\item{user}{username used to login to clowder}

\item{pass}{password}

\item{apikey}{optional used preferentially over username/pass}

\item{baseurl}{the url used to access clowder api e.g. clowder.ncsa.illinois.edu/clowder/api}
}
\description{
Setup authentication parameters for connecting to clowder
}
\details{
use clowder auth to save credentials for use with clowder functions
}
\examples{
\dontrun{
clowder_auth(user = "exampleuser@examplemail.com",
            pass = "pass",
            baseurl = "https://clowder.ncsa.illinois.edu/clowder/api")
}

}
