% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clowder_load_file.R
\name{clowder_load_file}
\alias{clowder_load_file}
\title{Load file from clowder uid into memory}
\usage{
clowder_load_file(uid = NULL)
}
\arguments{
\item{uid}{character string of the unique clowder uid for the file you want to load}
}
\value{
A data.table containing the file loaded from the uid
}
\description{
Load file from clowder uid into memory
}
\examples{
\dontrun{
clowder_load_file(uid ='5d78e1234567891011121f34')
}
}
